[core] Migrate `getMCPActionsForAgent` logic to `AgentMCPActionResource.listByAgent` and remove deprecated implementation
