+++ front/pages/api/w/[wId]/labs/mcp_actions/[agentId]/index.ts
@@ -9,7 +9,7 @@ import { getAgentConfiguration } from "@app/lib/api/assistant/configuration/agen
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
-import { AgentStepContentResource } from "@app/lib/resources/agent_step_content_resource";
+import { AgentMCPActionResource } from "@app/lib/resources/agent_mcp_action_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
 
@@ -106,14 +106,11 @@ async function handler(
         });
       }
 
-      const result = await AgentStepContentResource.getMCPActionsForAgent(
-        auth,
-        {
-          agentConfigurationId: agentConfiguration.sId,
-          limit,
-          cursor,
-        }
-      );
+      const result = await AgentMCPActionResource.listByAgent(auth, {
+        agentConfigurationId: agentConfiguration.sId,
+        limit,
+        cursor,
+      });
 
       if (result.isErr()) {
         const error = result.error;
